requirements/dev.txt
aiosmtplib==2.0.1
alembic==1.10.3
amqp==5.1.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-mail==1.2.7
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
humanize==4.6.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.10
outcome==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.4
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
starlette==0.26.1
tornado==6.3
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
wsproto==1.2.0
zipp==3.15.0
